+++ app.py
@@ -164,15 +164,7 @@ with gr.Blocks(title=init.ver, theme=gr.themes.Soft(
             
 ##################################################################################################################################            
             
-    with gr.Tab("[BETA-0.2]AI Detector"):
-        with gr.Row():
-            gr.Label("Detector threshold")
-            aid_slider = gr.Slider(
-                value=1.000,
-                label="Threshold (it is better not to change, it is left for tests)",
-                minimum=0.001,
-                maximum=1.001
-            ) 
+    with gr.Tab("AI Detector"):
         with gr.Row():
             gr.Label("Analyze image")
         with gr.Row():
@@ -248,8 +240,8 @@ with gr.Blocks(title=init.ver, theme=gr.themes.Soft(
     
     promptgen_button.click(multi.PromptGenetator, inputs=[prompt_input, pg_prompts, pg_max_length, randomize_temp], outputs=promptgen_output)
     
-    aid_single_button.click(multi.AiDetector_single, inputs=[aid_input_single, aid_slider], outputs=aid_output_single)
-    aid_batch_button.click(multi.AiDetector_batch, inputs=[aid_input_batch, aid_slider], outputs=aid_output_batch)
+    aid_single_button.click(multi.AiDetector_single, inputs=aid_input_single, outputs=aid_output_single)
+    aid_batch_button.click(multi.AiDetector_batch, inputs=aid_input_batch, outputs=aid_output_batch)
     aid_clear_button.click(multi.AID_Clear, outputs=aid_clearp)
     
     clear_all_button.click(init.clear_all, outputs=clear_all_tb)